backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-simple-captcha==0.5.18
django-redis==5.2.0
pillow==10.0.0
bleach==6.0.0
django-mptt==0.14.0
celery==5.3.4
redis==5.0.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.7
django-cors-headers==4.4.0
daphne==4.0.0